Typo and other tuneups for the gnocchi10 blog post.
Feel free to take what you like and ignore the rest. I mostly tried to
focus on issues with grammar, but there are minor structural changes
too.

The only other change I might suggest is to give a bit more of a hint
or clue that the hope is that all this fine stuff will address the
scaling issues described at the start. That point is introduced but
then not really addressed in any strong way later in the text.